+++ fightstreeter/leaderboards.py
@@ -2,15 +2,23 @@
 
 import sqlite3
 from datetime import datetime
+from zoneinfo import ZoneInfo
 
 import pandas as pd
+from flask import Blueprint, render_template
 
 from constants import (
     charid_map,
     fetch_league_name,
     fetch_mr_league_name,
     get_kudos_class,
+    phase_dates,
 )
+from fightstreeter import db
+
+from .awards import generate_awards
+
+bp = Blueprint("leaderboards", __name__, url_prefix="/leaderboards")
 
 
 def generate_leaderboards(
@@ -231,3 +239,79 @@ def generate_kudos_board(hs_df: pd.DataFrame) -> list[dict[str, str | int]]:
         )
 
     return kudos_leaders
+
+
+def get_list_of_dates():
+    """Returns a list of dates the site has data for"""
+    dates_with_data = []
+    with sqlite3.connect("cfn-stats.db") as conn:
+        cursor = conn.cursor()
+        sql = """SELECT date from last_update ORDER BY date DESC;"""
+        cursor.execute(sql)
+        results = cursor.fetchall()
+        dates_with_data = [x[0] for x in results]
+    return dates_with_data
+
+
+@bp.route("/", defaults={"date_req": ""})
+@bp.route("/<string:date_req>")
+def leaderboards(date_req: str) -> str:
+    """Displays MR/LP/Kudos leaderboards and stats for the club."""
+
+    # hit db to get latest data point
+    latest_data_date_sql = """SELECT date, download_complete, parsing_complete
+                            FROM last_update
+                            WHERE download_complete = 1 AND parsing_complete = 1
+                            ORDER BY date DESC
+                            LIMIT 1;"""
+
+    if not date_req:
+        try:
+            with sqlite3.connect("cfn-stats.db") as conn:
+                cursor = conn.cursor()
+                cursor.execute(latest_data_date_sql)
+                results = cursor.fetchone()
+                if results[0]:
+                    date_req = results[0]
+        except sqlite3.Error as e:
+            print(e)
+
+    split_date = date_req.split("-")
+    req_datetime = datetime.now(ZoneInfo("America/Los_Angeles"))
+    if len(split_date) == 3:
+        y, m, d = split_date
+        if len(y) == 4 and len(m) == 2 and len(d) == 2:
+            yint = int(y)
+            mint = int(m)
+            dint = int(d)
+            req_datetime = datetime.now(ZoneInfo("America/Los_Angeles")).replace(
+                microsecond=0,
+                second=0,
+                minute=0,
+                hour=12,
+                year=yint,
+                month=mint,
+                day=dint,
+            )
+
+    top_10_boards, top_10_grouped = generate_leaderboards(req_datetime)
+
+    awards_list = generate_awards()
+
+    date_list = get_list_of_dates()
+
+    final_list = []
+
+    for date in date_list:
+        for phases in phase_dates.items():
+            if phases[1][0] <= date <= phases[1][1]:
+                final_list.append((date, phases[0]))
+
+    return render_template(
+        "club_leaderboards.html.j2",
+        top_10_boards=top_10_boards,
+        top_10_grouped=top_10_grouped,
+        awards_list=awards_list,
+        date_selected=req_datetime.strftime("%Y-%m-%d"),
+        date_list=final_list,
+    )
